requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
python-dotenv==1.0.1
langchain==0.3.3
PyPDF2==3.0.1
chromadb==0.5.13
faiss-cpu==1.9.0
pdf2image==1.17.0
